requirements.txt
pandas
numpy
requests
sqlalchemy
python-dotenv
jupyter
openpyxl

